feat: support ssr generation
